refactor: refactor styles of main page components, users table

+++ src/helpers/convertDate.ts
@@ -2,7 +2,7 @@ export const convertDate = (date: string | number): string => {
   const newDate = new Date(date);
 
   const formatter = new Intl.DateTimeFormat('en', {
-    month: 'long',
+    month: 'short',
     day: 'numeric',
     hour: 'numeric',
     minute: 'numeric',
